Extracted Modal into button functionality, incorporated into UI
