+++ screens/UserSlotScreen.tsx
@@ -5,15 +5,12 @@ import SlotContainer from "../components/generic/SlotContainer"
 
 import { useStore } from '../Utilities/store';
 import { IsoDateToWeekDay,retrieveDatesFromChunks } from "../Utilities/slot_utilities"
-import EvaluationSlotPicker from "../components/complex/EvaluationSlotPicker"
 import { SortByDateAscending } from "../Utilities/slot_utilities"
 
 export default function UserSlotsScreen() {
 
     const initSlots = useStore((store) => store.initSlots);
     const slotChunks = useStore((store) => store.Slots);
-    const [visible, setVisible] = React.useState(false);
-    const [mainModalVisible, setMainModalVisible] = React.useState(false);
 
     let slotDays:string[] = [];// custom user facing strings
 
@@ -35,26 +32,10 @@ export default function UserSlotsScreen() {
         console.log("wtf")
     },[])
 
-    const bla = () => {
-        console.log("pressed fucking button")
-        setVisible(true);
-        setMainModalVisible(true)
-    }
 
-    const deactivate = () => {
-        console.log("pressed Cancel from other screen")
-        setVisible(false);
-        setMainModalVisible(false)
-    }
-    
 
     return (
         <SafeAreaView style={{flex: 1}}>
-            <Modal transparent={true} visible={mainModalVisible}>
-                <EvaluationSlotPicker modalVisible={mainModalVisible} onDismissModal={deactivate}></EvaluationSlotPicker>
-            </Modal>
-            <Button title="wowow" onPress={bla}>
-            </Button>
             <FlatList 
                 data={slotDays}
                 renderItem={({item}) =>(
